Add frontend builder using vite, use with cvss
Features coming down the line will require more involved JS. A bundler
will help organize and minify the JS, and also enable Typescript
checking. This adds a `frontend` directory, sets up an npm project in it
using vite, and move the CVSS code into it as an example.

Adds a container for the `local.yml` compose file that watches and
rebuilds the JS when changed, and an (UNTESTED) build process to the
production django dockerfile.