DRAFT: GQL in Django for Oplogs and Oplog Entries
A bunch of code to automatically generate an interface similar to what
Hasura provides for Django models, though it's not 1:1 yet.

The idea was to implement the oplog entry endpoints in Django so that
we could remove Hasura's observational triggers, since they also trigger
for Django saves that Django already knows about. We could then add the
Django GQL endpoint to Hasura as a remote endpoint, and have it mimic
the Hasura API for backwards compatibility.

However, Oplog also references its associated Project, which in turn
transitively associates with a whole slew of other models. To support
getting the project of the oplog, the Django GraphQL would have to
support returning a Project and all of its endpoints. At which point,
we may as well deprecate Hasura and provide a "v2" gql endpoint purely
in Django (which I would like to do eventually, but is a bigger
change).